fix set_simulation_rate_multiplier(), update tests, move load_scenario()
